agentic_wallet_poc/requirements.txt
pandas
numpy
scikit-learn
litellm
web3
eth-abi
eth-utils
pydantic
pytest
jupyter
python-dotenv
